fix(total): 전체 콘텐트 레이아웃 수정 #317
